feat(release): collapse harn/harn-lsp/harn-dap into one multi-call binary
harn-lsp and harn-dap each statically re-linked nearly the same dependency
closure as harn, so the release shipped three fat binaries (185 + 128 + 82 MB
uncompressed) and ran the expensive final LTO link three times per target.

Make harn-lsp and harn-dap thin lib+shim crates and link them into harn,
dispatching on argv[0]: when the binary is invoked as harn-lsp / harn-dap it
runs that subsystem, else the CLI. The release now builds ONE binary and ships
harn-lsp / harn-dap as symlinks to it (Unix) or copies (Windows, no dependable
unprivileged symlinks). install.sh and `harn upgrade` likewise place them as
on-disk symlinks so an install is one real binary plus two links, not three
copies.

Wins: ~210 MB less uncompressed per Unix target (tarball ~130 -> ~70 MB), the
slow LTO link runs once instead of three times, and editors that spawn
harn-lsp / harn-dap by path keep working unchanged. macOS now signs/notarizes
only harn; the symlinks resolve to it.
